requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
eventlet==0.36.1
python-dotenv==1.0.1
SQLAlchemy==2.0.32
pyepics==3.5.3
pydantic==2.8.2